+++ src/scanner.cpp
@@ -1,4 +1,4 @@
-#include "scanner.hpp"
+#include "../includes/scanner.hpp"
 
 int main(int argc, char **argv) {
     if(argc != 2) {
